[Ted] page 모듈 기능 수정 / news 모듈 기능 수정
1. page 모듈 기능 수정
- page.model.ts 에 ownerId 필드 추가
- updatePage.dto.ts 에 location 필드 삭제(location 업데이트 불가하도록 크리티컬수정)
- page.service.ts 내 updateOne 함수에 location 중복체크하던 로직 제거 / findById 로 변경하면서 page 존재여부, owner 권한 여부 체크 추가
- page.service.ts 내 deleteOne 함수에 ownerId 파라미터로 추가하면서 page 존재여부, owner 권한 여부 체크 추가
- 위 수정에 따라 page.service.spec.ts 수정
- 위 수정에 따라 page.controller.ts 수정

2. news 모듈 수정
- news.service.ts 내 updateOne 함수에 news 존재여부 체크 추가
- news.service.ts 내 deleteOne 함수에 news 존재여부 체크 추가
- 위 수정사항에 따라 news.service.spec.ts 수정

+++ src/news/news.service.ts
@@ -27,6 +27,8 @@ export class NewsService {
 
     async updateOne(newsId: string, updateNewsDTO: UpdateNewsDTO): Promise<NewsDocument | undefined> {
         const news = await this.newsModel.findById(newsId);
+        if (!news)
+            throw new NotFoundException('not found news');
         if (updateNewsDTO.ownerId.toString() !== news.ownerId)
             throw new UnauthorizedException('owner only can update!');
         delete updateNewsDTO.ownerId;
@@ -40,6 +42,8 @@ export class NewsService {
         // 임시방편으로 replace 문 적용
         const filteredNewsId = newsId.replace("\b", "");
         const news = await this.newsModel.findById(filteredNewsId);
+        if (!news)
+            throw new NotFoundException('not found news');
         if (ownerId.toString() !== news.ownerId)
             throw new UnauthorizedException('owner only can delete!');
         return this.newsModel.findByIdAndDelete(filteredNewsId);
